Update examples for typed Remo macros
