+++ examples/ios/README.md
@@ -102,4 +102,4 @@ RemoExample.xcworkspace
 └── RemoExampleUITests/           # UI automation tests
 ```
 
-Global capabilities (`items.add`, `items.remove`, `items.clear`, navigation, UI) are registered via `#remo` inline forms inside an `await #remo { ... }` block on the root `ContentView.task {}` — they live for the app's lifetime and are fully stripped from release builds. The Grid tab (`UIKitDemoViewController`) registers `grid.*` capabilities while it is alive and unregisters them on `deinit`. `AppStore.items` is pushed into the Grid tab via `UIKitDemoScreen.updateUIViewController` whenever the array changes.
+Global capabilities (`items.add`, `items.remove`, `items.clear`, navigation, UI) are registered as local typed `RemoCapability` enums inside a single `await #Remo { ... }` block on the root `ContentView.task {}`. The `await #remoScope { ... }` wrapper keeps them registered for the task lifetime and strips the whole island from release builds. The Grid tab (`UIKitDemoViewController`) follows the same pattern in UIKit: local typed capability enums live inside `#Remo { ... }`, while `#remoScope(scopedTo: self)` tracks them for unregister-on-disappear behavior. `AppStore.items` is pushed into the Grid tab via `UIKitDemoScreen.updateUIViewController` whenever the array changes.
